% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccdc_theme.R
\name{bccdc_theme}
\alias{bccdc_theme}
\title{Custom BCCDC ggplot theme}
\usage{
bccdc_theme(include_logo = TRUE)
}
\arguments{
\item{include_logo}{Optional boolean, if \code{TRUE} includes logo in bottom-left
corner of the plot}
}
\description{
Custom BCCDC ggplot theme
}
